[#2374] Refactor: move dashboard shared logic to shared folder
