+++ client/src/components/dashboard/shared/DashboardViewer.scss
@@ -1,4 +1,4 @@
-@import "../../styles/modules/importAll";
+@import "../../../styles/modules/importAll";
 
 html, body, #root {
     min-width: initial !important;
