+++ client/src/components/dashboard/shared/DashboardViewerItem.scss
@@ -1,4 +1,4 @@
-@import "../../styles/modules/importAll";
+@import "../../../styles/modules/importAll";
 
 .DashboardCanvasItem {
   position: relative;
